+++ kyoukai/app.py
@@ -15,7 +15,7 @@ from werkzeug.wrappers import Request, Response
 from kyoukai.asphalt import HTTPRequestContext
 from kyoukai.blueprint import Blueprint
 
-__version__ = "2.2.0"
+__version__ = "2.2.1"
 
 logger = logging.getLogger("Kyoukai")